Avro: support writing SMALLINT to Avro
The Postgres/GPDB type SMALLINT has no corresponding Avro type[0].

In this commit we make SMALLINT types being written to Avro external
sources appear as Avro integer type. For reading Avro data, the
Greenplum external table should always use INT or BIGINT.

[0] https://avro.apache.org/docs/current/spec.html#schema_primitive

Authored-by: Oliver Albertini <oalbertini@vmware.com>

+++ server/pxf-hdfs/src/test/java/org/greenplum/pxf/plugins/hdfs/AvroResolverTest.java
@@ -67,7 +67,7 @@ public class AvroResolverTest {
         assertEquals(false, genericRecord.get(0));
         assertEquals(ByteBuffer.wrap(new byte[]{(byte) 49}), genericRecord.get(1));
         assertEquals(23456789L, genericRecord.get(2));
-        assertEquals((short) 1, genericRecord.get(3));
+        assertEquals(1, genericRecord.get(3));
         assertEquals((float) 7.7, genericRecord.get(4));
         assertEquals(6.0, genericRecord.get(5));
         assertEquals("row1", genericRecord.get(6));
